project/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
mysql-connector-python==8.3.0
pydantic==2.6.1
python-multipart==0.0.9